Put the call helpers where they are used
Calls.ts was scaffolding left from making connections numeric: two functions
with one caller each and nothing in common but the word call.

isCall is the line between what the algorithm plans with and what it ignores,
which is a decision Normalise makes, so it lives there. callAt undoes the route
numbering to find a stop time in the feed, which is what Connection is for, and
nothing outside it needs it so it is no longer exported.

They still share one definition of a call, because a filter and a count that
disagreed would resolve connections to the wrong stop times.
